+++ src/types/logger.d.ts
@@ -1,5 +1,7 @@
 import { PluginFunction } from 'vue'
 
+export type LogLevel = 'debug' | 'info' | 'warn' | 'error' | 'log'
+
 /**
  * @class
  * @example
@@ -77,7 +79,7 @@ export declare class VueLogger {
   /**
    * Returns the currently applied logging level (one of: debug, info, warn, error, log).
    */
-  level (): string
+  level (): LogLevel
   install: PluginFunction<LoggerOptions>
   static install: PluginFunction<LoggerOptions>
 }
@@ -98,7 +100,17 @@ export interface LoggerOptions {
   /**
    * @field level {string} the logging level (one of: debug, info, warn, error)
    */
-  level?: string
+  level?: LogLevel
+  /**
+   * @field callerInfo {boolean} whether information about the caller function should be included
+   * @see {@link CallerInfo}
+   */
+   callerInfo?: boolean
+   /**
+    * @field provide a custom formatted string for the log message prefix (preceeds the log arguments)
+    * @param event {@link LogEvent}
+    */
+   prefixFormat? (event: Readonly<Pick<LogEvent, 'level' | 'caller'>>): string
   /**
    * @field beforeHooks {LoggerHook[]} hooks invoked before a statement is logged, can be used to alter log arguments (use carefully)
    */
@@ -109,13 +121,43 @@ export interface LoggerOptions {
   afterHooks?: LoggerHook[]
 }
 
+/**
+ * Provides custom logic to be executed during a logging event.
+ *
+ * @interface
+ * @example
+ * const customHook: LoggerHook = {
+ *  run (event: LogEvent) {
+ *    // event.level is the log level invoked (e.g. debug, info, warn, error)
+ *    // event.argumentArray is the array of arguments which were passed to the logging method
+ *    // event.caller contains the caller function information (e.g. function name, file name, line number), if LoggerOptions.callerInfo is true
+ *  }
+ * }
+ */
 export interface LoggerHook {
   run (event: LogEvent): void
   install? (options: LoggerOptions): void
   props?: { [key: string]: any }
 }
 
+/**
+ * Contains information for a log invocation (log level, array of arguments, caller function info).
+ *
+ * @interface
+ */
 export interface LogEvent {
-  level: string
+  level: LogLevel
   argumentArray: any[]
+  caller?: CallerInfo
+}
+
+/**
+ * Information about the caller function which invoked the logger.
+ *
+ * @interface
+ */
+export interface CallerInfo {
+  fileName?: string
+  functionName?: string
+  lineNumber?: string
 }