+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # vue-logger-plugin changelog
 
+## [1.2.1] - 2022-02-09
+
+* Add caller function information support for Firefox browser
+* Update default value for prefixFormat option
+
+## [1.2.0] - 2022-02-05
+
+* Add caller function information, controlled via `callerInfo` option
+* Add custom message prefix formatting, controlled via `prefixFormat` option
+* Add LogLevel type
+* Replace tslint with eslint
+
 ## [1.1.3] - 2022-02-05
 
 * Bump tmpl from 1.0.4 to 1.0.5
@@ -34,6 +46,8 @@
     * built-in argument stringify hooks
     * on-demand log level adjustment and enable/disable via `apply`
 
+[1.2.1]:  https://github.com/dev-tavern/vue-logger-plugin/compare/v1.2.0...v1.2.1
+[1.2.0]:  https://github.com/dev-tavern/vue-logger-plugin/compare/v1.1.3...v1.2.0
 [1.1.3]:  https://github.com/dev-tavern/vue-logger-plugin/compare/v1.1.2...v1.1.3
 [1.1.2]:  https://github.com/dev-tavern/vue-logger-plugin/compare/v1.1.1...v1.1.2
 [1.1.1]:  https://github.com/dev-tavern/vue-logger-plugin/compare/v1.1.0...v1.1.1